// src/lib/api/types/categoryDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */

export interface CategoryDTO {
  /** Created timestamp */
  createdAt?: string;
  /** Category description */
  description?: string;
  /** Category ID */
  id: string;
  /** Whether category is active */
  isActive: boolean;
  /** Category name */
  name: string;
  /** Parent category ID */
  parentCategoryId?: string;
  /** Category slug */
  slug: string;
  /** Updated timestamp */
  updatedAt?: string;
}